+++ Release-Notes.md
@@ -1,5 +1,11 @@
 # Grüner Pass (ValidationCore-iOS) - Release Notes
 
+## v.2.0.2 (08.2021)
+
+- EPIEMSCO-1506: Pass information in completion handler if signed data was updated
+- EEPIEMSCO-1472: Improve keychain handling on iOS 12 to fix double refresh
+- other bugfixes
+
 ## v.2.0.0 (08.2021)
 
 - EPIEMSCO-1343: Integration von Time Server zum Abruf der aktuellsten Zeit für 3-G Statusanzeige